// apps/web/src/domains/creator/studio-cuttoon-editor/studio-cuttoon-stage-pointers-api.ts
export type StudioCuttoonStagePointersApi = {
  tryStageDownArmedTools: (e: any, stagePointerEvent: PointerEvent) => boolean;
  tryStageDownPixel: (e: any, stagePointerEvent: PointerEvent) => boolean;
  tryStageDownDraw: (e: any, stagePointerEvent: PointerEvent) => void;
  onStageDown: (e: any) => void;
  updateHealCloneCursorNodes: (destNorm: any, frame: any) => void;
  updateHistoryBrushCursorNode: (destNorm: any, frame: any) => void;
  snapPointToAdvancedRuler: (...args: any[]) => any;
  collectStrokeObjectSnapTargets: (excludeId?: string | null) => any;
  strokeObjectSnapTargetsFor: (strokeId: string, excludeId?: string | null) => any;
  applyStrokeObjectSnapToPoint: (x: number, y: number, options?: any) => { x: number; y: number };
  updateActiveShapeEndpoint: (...args: any[]) => boolean;
  onStageMove: (e: any) => void;
  appendFixedRateStrokeSamples: (...args: any[]) => void;
  appendFreehandStrokePoint: (...args: any[]) => void;
  appendDrawingCrdtSampleSuffix: (...args: any[]) => void;
  consumeFreehandPointerBatch: (...args: any[]) => boolean;
  publishAuthoritativeFreehandSuffix: (startSample: number) => any;
  queueStudioBg3dMagicFilterMaskPublication: (input: any) => void;
  queueStudioRasterDrawPromotion: (input: any) => void;
  releaseEndpointPointerSample: (...args: any[]) => any;
  studioPageElementsFromHistory: (pageId: string) => any;
  withStudioHokusaiSource: (...args: any[]) => any;
  rejectStudioHokusaiFailedStroke: (...args: any[]) => void;
  completeStudioLivingInkRejectedNoop: (...args: any[]) => void;
  rejectStudioLivingInkFailedStroke: (...args: any[]) => void;
  finishStudioLivingInkStroke: (...args: any[]) => Promise<void>;
  finishStudioHokusaiLiveStroke: (...args: any[]) => Promise<void>;
  sealStudioDrawReleaseInput: (...args: any[]) => any;
  finishStudioSpecialistStroke: (...args: any[]) => any;
  finishDrawingPointer: (...args: any[]) => void;
  onStagePointerCancel: (e: any) => void;
  onStageUp: (e: any) => void;
  drawBrushCursorLayer: (deferToFrame: boolean) => void;
  hideStrokeGuide: (deferToFrame?: boolean) => void;
  hideBrushCursorVisual: (deferToFrame?: boolean) => void;
  hideBrushCursor: (deferToFrame?: boolean) => void;
  updateStrokeGuide: (...args: any[]) => void;
  updateBrushCursor: (...args: any[]) => void;
  hideSmudgeCursor: () => void;
  hideHealCloneCursors: () => void;
  hideHistoryBrushCursor: () => void;
  hideLayerMaskCursor: () => void;
  hideFilterMaskCursor: () => void;
  restoreGroupDragPreview: (...args: any[]) => void;
  cancelCanvasGroupDrag: () => boolean;
  liveCanvasElementRect: (...args: any[]) => any;
  onStageDragMove: (e: any) => void;
  onStageDragEnd: () => void;
};